Added UserStatus check to gateway
### What's done:
 * Added status field to SaveUserDetails
 * Forbade most endpoints to be used by inactive users
