+++ api-gateway/src/main/kotlin/com/saveourtool/save/gateway/security/WebSecurityConfig.kt
@@ -12,11 +12,14 @@ import com.saveourtool.save.v1
 
 import org.springframework.context.annotation.Bean
 import org.springframework.core.annotation.Order
+import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpStatus
 import org.springframework.security.authentication.UserDetailsRepositoryReactiveAuthenticationManager
+import org.springframework.security.authorization.AuthenticatedReactiveAuthorizationManager
 import org.springframework.security.authorization.AuthorizationDecision
 import org.springframework.security.config.annotation.web.reactive.EnableWebFluxSecurity
 import org.springframework.security.config.web.server.ServerHttpSecurity
+import org.springframework.security.core.Authentication
 import org.springframework.security.core.userdetails.UserDetails
 import org.springframework.security.crypto.factory.PasswordEncoderFactories
 import org.springframework.security.crypto.password.PasswordEncoder
@@ -26,11 +29,13 @@ import org.springframework.security.web.server.authentication.HttpStatusServerEn
 import org.springframework.security.web.server.authentication.RedirectServerAuthenticationFailureHandler
 import org.springframework.security.web.server.authentication.RedirectServerAuthenticationSuccessHandler
 import org.springframework.security.web.server.authentication.logout.HttpStatusReturningServerLogoutSuccessHandler
+import org.springframework.security.web.server.authorization.AuthorizationContext
 import org.springframework.security.web.server.util.matcher.AndServerWebExchangeMatcher
 import org.springframework.security.web.server.util.matcher.NegatedServerWebExchangeMatcher
 import org.springframework.security.web.server.util.matcher.ServerWebExchangeMatcher
 import org.springframework.security.web.server.util.matcher.ServerWebExchangeMatcher.MatchResult
 import org.springframework.security.web.server.util.matcher.ServerWebExchangeMatchers
+import reactor.core.publisher.Mono
 import reactor.kotlin.core.publisher.cast
 
 /**
@@ -78,7 +83,7 @@ class WebSecurityConfig(
                 .pathMatchers(*allowedForInactiveEndpoints.toTypedArray()).access(::defaultAuthorizationDecision)
                 .pathMatchers("/api/**").access { authorization, authorizationContext ->
                     authorization.flatMap { backendService.findByName(it.name) }
-                        .filter { it.isActive() }
+                        .filter { it.isEnabled }
                         .flatMap { defaultAuthorizationDecision(authorization, authorizationContext) }
                         .defaultIfEmpty(AuthorizationDecision(false))
                 }
@@ -164,3 +169,26 @@ class WebSecurityConfig(
  */
 @Bean
 fun passwordEncoder(): PasswordEncoder = PasswordEncoderFactories.createDelegatingPasswordEncoder()
+
+/**
+ * Get default [AuthorizationDecision] by [authentication] and [authorizationContext]
+ *
+ * @param authentication
+ * @param authorizationContext
+ * @return [Mono] of [AuthorizationDecision]
+ */
+fun defaultAuthorizationDecision(
+    authentication: Mono<Authentication>,
+    authorizationContext: AuthorizationContext,
+): Mono<AuthorizationDecision> = AuthenticatedReactiveAuthorizationManager.authenticated<AuthorizationContext>().check(
+    authentication, authorizationContext
+).map {
+    if (!it.isGranted) {
+        // if request is not authorized by configured authorization manager, then we allow only requests w/o Authorization header
+        // then backend will return 401, if endpoint is protected for anonymous access
+        val hasAuthorizationHeader = authorizationContext.exchange.request.headers[HttpHeaders.AUTHORIZATION].isNullOrEmpty()
+        AuthorizationDecision(hasAuthorizationHeader)
+    } else {
+        it
+    }
+}
